requirements.txt
discord.py>=2.0.0
requests>=2.28.1
google-generativeai>=0.2.0
python-dotenv>=0.20.0
